(closes #5) Add router
